Add conditional static linking to configure script
glibc is bad. Especially at static linking. So we detect it by trying to
include a glibc-specific header (__GLIBC__ macro is unreliable, uClibc
defines it and probably others), and disable static linking if it succeeds

musl-clang and musl-gcc are used in preference to normal compilers, as
this will allow us to statically link to musl even on glibc-native systems